Reorganize IO things into `io` submodules
